refactor: move ImportExport.js file and add boilerplate files
